Add shading capability to grid
Added functionality touches to Shader button, and now have
dynamic shading added.

Part of the issue was that I had the shading function called within my
main event delegation with the other drawing functions. This caused it
to either "not work" or it became a memory leak that would crash slow
computers. It is now currently attached purely to the related button,
and nowhere else.