Fix opacity shading bug
The shader opacity would conflict with the normal pen, disallowing
it to draw over the cell, visually.

To correct this, I simply set it so that when you draw with the regular
pen, it would set the opacity back to 1. The bug happened because the
opacity would stay (as intended) lower than the baseline.